move flow group
